feat: replace m*pbox geocoding with google maps
Lots of stuff going on... added support for Google Maps links, updated TODO, etc.

Mapbox's results were inconsistent, and google maps links offer a nice user confirmation. By going with Google Maps for geocoding, we have better piece-of-mind that the results of the geocoding query and the google maps link will be telling the same story.